more problems as usuall

+++ ExpandoFeedTransformer/Program.cs
@@ -8,7 +8,7 @@ namespace ExpandoFeedTransformer
     internal class Program
     {
         private const string path = "\\\\AzetCool-Pohoda\\POHODA_SK_E1_DATA\\Dokumenty\\ACecom\\Obrázky\\";
-        private static uint num = 3;
+        private static ulong num = 3;
 
         private static async Task Main(string[] args)
         {
@@ -24,7 +24,7 @@ namespace ExpandoFeedTransformer
                 Console.WriteLine("Exception: " + e.Message);
             }
 
-            var v = uint.Parse($"3{DateTime.Now:yyMMdd}{line}");
+            var v = ulong.Parse($"3{DateTime.Now:yyMMdd}{line}");
             num = v;
             Console.WriteLine("Starting pohoda mServer");
             var mServer = new PohodaMServer("test", "\"C:\\Program Files (x86)\\STORMWARE\\POHODA SK E1\"",
